Update CLI tool to net6 (LTS) and net7, drop netcore2.1
netcore2.1 is long-since deprecated, so dropping support for it.
Instead, adding the current LTS and STS versions.  As a policy going
forward, we should maintain these targets according to their lifecycles.
The dotnet lifecycle is fairly straight-forward (LTS = 3 years, STS =
1.5 years).

Technically, this is a breaking change, as a machine configured with sdk
<net6.0 might not be able to run the tool anymore.